Move tests to *.spec.js (#468)
* Move tests to *.spec.js

* Update .codeclimate.yml
